+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log for ~~AzureCP~~ EntraCP
 
+## EntraCP v27.0 - Unreleased
+
+* Ensure that all group members are retrieved when only users members of specified groups can be found in SharePoint
+* Update the script that provisions tenant with test users and groups, to be more reliable and provision 999 users (instead of 50), so tests are more realistics
+* Improve tests
+
 ## EntraCP v26.0.20240627.35 enhancements & bug-fixes - Published in June 27, 2024
 
 * Fix an NullReferenceException in a very rare scenario where ClaimsPrincipal.Identity is null